PythonFunctions/requirements.txt
azure-functions==1.0.0a5
azure-functions-worker==1.0.0a6
grpcio==1.14.2
grpcio-tools==1.14.2
protobuf==3.6.1
six==1.11.0
beautifulsoup4==4.6.3
requests
